complete Stacked chart component
