+++ src/components/Charts/Stacked.jsx
@@ -1,7 +1,40 @@
 import React from "react";
+import {
+  ChartComponent,
+  SeriesDirective,
+  SeriesCollectionDirective,
+  Inject,
+  Legend,
+  Category,
+  StackingColumnSeries,
+  Tooltip,
+} from "@syncfusion/ej2-react-charts";
+import {
+  stackedCustomSeries,
+  stackedPrimaryYAxis,
+  stackedPrimaryXAxis,
+} from "../../data/dummyData";
 
-const Stacked = () => {
-  return <div>Stacked Chart</div>;
+const Stacked = ({ width, height }) => {
+  return (
+    <ChartComponent
+      id="charts"
+      width={width}
+      height={height}
+      primaryXAxis={stackedPrimaryXAxis}
+      primaryYAxis={stackedPrimaryYAxis}
+      chartArea={{ border: { width: 0 } }}
+      tooltip={{ enable: true }}
+      legendSettings={{ background: "white" }}
+    >
+      <Inject services={[Legend, Category, StackingColumnSeries, Tooltip]} />
+      <SeriesCollectionDirective>
+        {stackedCustomSeries.map((item, index) => (
+          <SeriesDirective key={index} {...item} />
+        ))}
+      </SeriesCollectionDirective>
+    </ChartComponent>
+  );
 };
 
 export default Stacked;
